GEOJSON_MAP_VIEW/requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
pandas==2.2.3
duckdb==1.4.4
sentence-transformers==5.2.3
faiss-cpu==1.13.2
streamlit==1.44.1
pydeck==0.9.1
geopy==2.4.1
openai==1.75.0
openpyxl==3.1.5
shapely==2.0.7
pyproj==3.7.0
pytest==8.3.5
